Revert "Cleanup domain objects"
This reverts commit 66e7698a
