+++ save-cloud-common/src/commonMain/kotlin/org/cqfn/save/entities/Project.kt
@@ -1,0 +1,34 @@
+package org.cqfn.save.entities
+
+import org.cqfn.save.utils.EnumType
+
+import kotlinx.serialization.Serializable
+
+/**
+ * @property owner
+ * @property name
+ * @property url
+ * @property description description of the project, may be absent
+ * @property status status of project
+ * @property public
+ * @property user the user that has created this project
+ * @property adminIds comma-separated list of IDs of users that are admins of this project
+ */
+@Entity
+@Serializable
+data class Project(
+    var owner: String,
+    var name: String,
+    var url: String?,
+    var description: String?,
+    @Enumerated(EnumType.STRING)
+    var status: ProjectStatus,
+    var public: Boolean = true,
+) {
+    /**
+     * id of project
+     */
+    @Id
+    @GeneratedValue
+    var id: Long? = null
+}
