--- save-cloud-common/src/commonMain/kotlin/org/cqfn/save/entities/ProjectDto.kt
@@ -1,25 +1,0 @@
-package org.cqfn.save.entities
-
-import org.cqfn.save.utils.EnumType
-
-import kotlinx.serialization.Serializable
-
-/**
- * @property owner
- * @property name
- * @property url
- * @property description description of the project, may be absent
- * @property status status of project
- * @property public
- */
-@Serializable
-data class ProjectDto(
-    val id: Long,
-    val owner: String,
-    val name: String,
-    val url: String?,
-    val description: String?,
-    val status: ProjectStatus,
-    val public: Boolean = true,
-    val username: String,
-)
