+++ save-cloud-common/src/commonMain/kotlin/org/cqfn/save/testsuite/TestSuiteDto.kt
@@ -1,6 +1,6 @@
 package org.cqfn.save.testsuite
 
-import org.cqfn.save.entities.ProjectDto
+import org.cqfn.save.entities.Project
 
 import kotlinx.serialization.Serializable
 
@@ -18,7 +18,7 @@ data class TestSuiteDto(
     val type: TestSuiteType?,
     val name: String,
     val description: String?,
-    val project: ProjectDto? = null,
+    val project: Project? = null,
     val testRootPath: String,
     val testSuiteRepoUrl: String? = null,
     val language: String? = null,
