+++ save-frontend/src/main/kotlin/org/cqfn/save/frontend/components/basic/HtmlElements.kt
@@ -4,7 +4,7 @@
 
 package org.cqfn.save.frontend.components.basic
 
-import org.cqfn.save.entities.ProjectDto
+import org.cqfn.save.entities.Project
 
 import csstype.BorderRadius
 import react.CSSProperties
@@ -14,7 +14,7 @@ import react.dom.span
 /**
  * @param project
  */
-fun RBuilder.privacySpan(project: ProjectDto) {
+fun RBuilder.privacySpan(project: Project) {
     span("border ml-2 pr-1 pl-1 text-xs text-muted ") {
         attrs["style"] = kotlinext.js.jsObject<CSSProperties> {
             borderRadius = "2em".unsafeCast<BorderRadius>()
