+++ save-frontend/src/main/kotlin/org/cqfn/save/frontend/components/views/CreationView.kt
@@ -9,7 +9,7 @@ package org.cqfn.save.frontend.components.views
 
 import org.cqfn.save.entities.GitDto
 import org.cqfn.save.entities.NewProjectDto
-import org.cqfn.save.entities.ProjectDto
+import org.cqfn.save.entities.Project
 import org.cqfn.save.entities.ProjectStatus
 import org.cqfn.save.frontend.components.basic.InputTypes
 import org.cqfn.save.frontend.components.basic.inputTextFormOptional
@@ -171,8 +171,7 @@ class CreationView : AbstractView<ProjectSaveViewProps, ProjectSaveViewState>(tr
             return
         }
         val newProjectRequest = NewProjectDto(
-            ProjectDto(
-                id = -1,
+            Project(
                 fieldsMap[InputTypes.OWNER]!!.trim(),
                 fieldsMap[InputTypes.PROJECT_NAME]!!.trim(),
                 fieldsMap[InputTypes.PROJECT_URL]?.trim(),
