+++ save-frontend/src/main/kotlin/org/cqfn/save/frontend/components/views/ProjectView.kt
@@ -186,7 +186,7 @@ class ProjectView : AbstractView<ProjectExecutionRouteProps, ProjectViewState>(f
     private var standardTestSuites: List<TestSuiteDto> = emptyList()
     private val selectedStandardSuites: MutableList<String> = mutableListOf()
     private var gitDto: GitDto? = null
-    private var project = ProjectDto(1, "N/A", "N/A", "N/A", "N/A", ProjectStatus.CREATED, username = "N/A")
+    private var project = Project("N/A", "N/A", "N/A", "N/A", ProjectStatus.CREATED)
     private val projectInformation = mutableMapOf(
         "Tested tool name: " to "",
         "Description: " to "",
